1. Files moved for MVVM creating; 2. ViewModel creating started.
