--- app/src/main/java/com/mtsteta/homework/data/dto/CategoryDto.kt
@@ -1,5 +1,0 @@
-package com.mtsteta.homework.data.dto
-
-data class CategoryDto (
-    val name: String
-)
