--- app/src/main/java/com/mtsteta/homework/data/features/categories/CategoriesDataSource.kt
@@ -1,7 +1,0 @@
-package com.mtsteta.homework.data.features.categories
-
-import com.mtsteta.homework.data.dto.CategoryDto
-
-interface CategoriesDataSource {
-    fun getCategories(): List<CategoryDto>
-}
