--- app/src/main/java/com/mtsteta/homework/data/features/movies/MoviesDataSource.kt
@@ -1,7 +1,0 @@
-package com.mtsteta.homework.data.features.movies
-
-import com.mtsteta.homework.data.dto.MovieDto
-
-interface MoviesDataSource {
-    fun getMovies(): List<MovieDto>
-}
