+++ app/src/main/java/com/mtsteta/homework/model/dto/CategoryDto.kt
@@ -1,0 +1,5 @@
+package com.mtsteta.homework.model.dto
+
+data class CategoryDto (
+    val name: String
+)
