+++ app/src/main/java/com/mtsteta/homework/model/features/categories/CategoriesDataSource.kt
@@ -1,0 +1,7 @@
+package com.mtsteta.homework.model.features.categories
+
+import com.mtsteta.homework.model.dto.CategoryDto
+
+interface CategoriesDataSource {
+    fun getCategories(): List<CategoryDto>
+}
