+++ app/src/main/java/com/mtsteta/homework/model/features/movies/MoviesDataSource.kt
@@ -1,0 +1,7 @@
+package com.mtsteta.homework.model.features.movies
+
+import com.mtsteta.homework.model.dto.MovieDto
+
+interface MoviesDataSource {
+    fun getMovies(): List<MovieDto>
+}
