+++ app/src/main/java/com/mtsteta/homework/model/models/CategoriesModel.kt
@@ -1,0 +1,7 @@
+package com.mtsteta.homework.model.models
+
+import com.mtsteta.homework.model.features.categories.CategoriesDataSource
+
+class CategoriesModel(private val categoriesDataSource: CategoriesDataSource) {
+    fun getCategories() = categoriesDataSource.getCategories()
+}
