+++ app/src/main/java/com/mtsteta/homework/model/models/MoviesModel.kt
@@ -1,0 +1,7 @@
+package com.mtsteta.homework.model.models
+
+import com.mtsteta.homework.model.features.movies.MoviesDataSource
+
+class MoviesModel(private val moviesDataSource: MoviesDataSource) {
+    fun getMovies() = moviesDataSource.getMovies()
+}
