--- app/src/main/java/com/mtsteta/homework/presentation/models/CategoriesModel.kt
@@ -1,7 +1,0 @@
-package com.mtsteta.homework.presentation.models
-
-import com.mtsteta.homework.data.features.categories.CategoriesDataSource
-
-class CategoriesModel(private val categoriesDataSource: CategoriesDataSource) {
-    fun getCategories() = categoriesDataSource.getCategories()
-}
