--- app/src/main/java/com/mtsteta/homework/presentation/models/MoviesModel.kt
@@ -1,7 +1,0 @@
-package com.mtsteta.homework.presentation.models
-
-import com.mtsteta.homework.data.features.movies.MoviesDataSource
-
-class MoviesModel(private val moviesDataSource: MoviesDataSource) {
-    fun getMovies() = moviesDataSource.getMovies()
-}
