--- app/src/main/java/com/mtsteta/homework/presentation/recyclerviews/decorations/BottomSpaceItemDecoration.kt
@@ -1,14 +1,0 @@
-package com.mtsteta.homework.presentation.recyclerviews.decorations
-
-import android.graphics.Rect
-import android.view.View
-import androidx.recyclerview.widget.RecyclerView
-
-class BottomSpaceItemDecoration(private val spaceSize: Int) : RecyclerView.ItemDecoration() {
-    override fun getItemOffsets(
-        outRect: Rect, view: View, parent: RecyclerView,
-        state: RecyclerView.State
-    ) {
-        outRect.bottom = spaceSize
-    }
-}
