--- app/src/main/java/com/mtsteta/homework/presentation/recyclerviews/viewholders/CategoriesViewHolder.kt
@@ -1,22 +1,0 @@
-package com.mtsteta.homework.presentation.recyclerviews.viewholders
-
-import android.view.View
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.mtsteta.homework.R
-import com.mtsteta.homework.data.dto.CategoryDto
-
-class CategoriesViewHolder(itemView: View, onCategoryItemClick: (title: String) -> Unit) : RecyclerView.ViewHolder(itemView) {
-    private val movieCategoryTextView: TextView = itemView.findViewById(R.id.textItemMovieCategory)
-
-    lateinit var item: CategoryDto
-
-    init {
-        movieCategoryTextView.setOnClickListener { onCategoryItemClick.invoke(item.name) }
-    }
-
-    fun bind(categoryDto: CategoryDto) {
-        item = categoryDto
-        movieCategoryTextView.text = categoryDto.name
-    }
-}
