--- app/src/main/java/com/mtsteta/homework/presentation/recyclerviews/viewholders/EmptyMoviesListViewHolder.kt
@@ -1,15 +1,0 @@
-package com.mtsteta.homework.presentation.recyclerviews.viewholders
-
-import android.view.View
-import android.widget.TextView
-import androidx.recyclerview.widget.RecyclerView
-import com.mtsteta.homework.R
-
-class EmptyMoviesListViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
-    private val emptyMoviesListTextView: TextView =
-        itemView.findViewById(R.id.textItemEmptyMoviesList)
-
-    fun bind() {
-        emptyMoviesListTextView.text = itemView.context.getString(R.string.empty_movies_list)
-    }
-}
