+++ app/src/main/java/com/mtsteta/homework/view/MainActivity.kt
@@ -1,0 +1,28 @@
+package com.mtsteta.homework.view
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.navigation.fragment.findNavController
+import androidx.navigation.ui.AppBarConfiguration
+import androidx.navigation.ui.setupActionBarWithNavController
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.mtsteta.homework.R
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_main)
+
+        setupBottomNavigation()
+    }
+
+    private fun setupBottomNavigation() {
+        val navView: BottomNavigationView = findViewById(R.id.nav_view)
+
+        val navController = supportFragmentManager.findFragmentById(R.id.nav_host_fragment)
+            ?.findNavController()
+
+        navView.setupWithNavController(navController!!)
+    }
+}
