+++ app/src/main/java/com/mtsteta/homework/view/recyclerviews/decorations/RightSpaceItemDecoration.kt
@@ -1,0 +1,14 @@
+package com.mtsteta.homework.view.recyclerviews.decorations
+
+import android.graphics.Rect
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class RightSpaceItemDecoration(private val spaceSize: Int) : RecyclerView.ItemDecoration() {
+    override fun getItemOffsets(
+        outRect: Rect, view: View, parent: RecyclerView,
+        state: RecyclerView.State
+    ) {
+        outRect.right = spaceSize
+    }
+}
