+++ app/src/main/java/com/mtsteta/homework/viewmodel/HomeViewModel.kt
@@ -1,0 +1,4 @@
+package com.mtsteta.homework.viewmodel
+
+class HomeViewModel {
+}
