+++ app/src/main/java/com/mtsteta/homework/viewmodel/ProfileViewModel.kt
@@ -1,0 +1,4 @@
+package com.mtsteta.homework.viewmodel
+
+class ProfileViewModel {
+}
